Load default KeyFactory instead of BC's implementation
BouncyCastle has a shortcut for getting its own KeyFactory implementations. Since some of them are removed, e.g. KeyFactory.RSA, we need to load other available implementations.In particular BKS KeyStore, that is only available in BouncyCastle, uses KeyFactory.RSA, so this Android modifications are necessary for the KeyStore to keep functioning.

Test: cts -m CtsLibcoreOkHttpTestCases
Bug: 67761667
Change-Id: I98adb4d2f475e8b600a44b655ade83cd94c1efbd